+++ src/app/commands/afk/AFK.ts
@@ -7,7 +7,7 @@ import ValidationError from '../../errors/ValidationError';
 import ServiceUtils from '../../utils/ServiceUtils';
 import ToggleAFK from '../../service/AFK/ToggleAFK';
 import discordServerIds from '../../service/constants/discordServerIds';
-import { LogUtils } from '../../utils/Log';
+import Log, { LogUtils } from '../../utils/Log';
 
 export default class AFK extends SlashCommand {
 	constructor(creator: SlashCreator) {
@@ -29,30 +29,34 @@ export default class AFK extends SlashCommand {
 		const { guildMember } = await ServiceUtils.getGuildAndMember(ctx);
 		const AFKRole = ServiceUtils.getAFKRole(guildMember.guild.roles);
 		if (!AFKRole) {
-			return ctx.send('AFK Role does not exist on this server');
+			return ctx.send({ content: 'AFK Role does not exist on this server', ephemeral: true });
 		}
 		let command: Promise<any>;
 		try {
 			const isAFK : boolean = await ToggleAFK(guildMember);
 			if (isAFK) {
-				return ctx.send(`${ctx.user.username} has gone AFK!`);
+				command = ctx.send(`${ctx.user.username} has gone AFK!`).catch(Log.error);
 			} else {
-				return ctx.send(`Welcome back ${ctx.user.username}!`);
+				command = ctx.send(`Welcome back ${ctx.user.username}!`).catch(Log.error);
 			}
 		} catch (e) {
 			this.handleCommandError(ctx, command);
 		}
 	}
 
-	handleCommandError(ctx: CommandContext, command: Promise<any>): void {
+	handleCommandError(ctx: CommandContext, command?: Promise<any>): void {
+		if (command == null) {
+			Log.error('command null for AFK');
+			return;
+		}
 		command.then(() => {
 			return ctx.send(`${ctx.user.username} Sent you a DM with information.`);
 		}).catch(e => {
 			if (e instanceof ValidationError) {
-				return ctx.send(e.message);
+				return ctx.send({ content: `${e.message}`, ephemeral: true });
 			} else {
 				LogUtils.logError('failed to handle AFK command', e);
-				return ctx.send('Sorry something is not working and our devs are looking into it.');
+				return ServiceUtils.sendOutErrorMessage(ctx);
 			}
 		});
 	}