+++ src/app/events/MessageCreate.ts
@@ -1,6 +1,4 @@
 import messageCreateOnBountyBoard from './bounty/MessageCreateOnBountyBoard';
-import messageSetScoapRoles from './scoap-squad/messageSetScoapRoles';
-import messageLaunchFirstQuest from './first-quest/messageLaunchFirstQuest';
 import { Message } from 'discord.js';
 import { DiscordEvent } from '../types/discord/DiscordEvent';
 import MessageCreateOnDEGEN from './chat/MessageCreateOnDEGEN';
@@ -15,24 +13,20 @@ export default class implements DiscordEvent {
 		try {
 			if(message.author.bot) return;
 			// DEGEN says hello
-			MessageCreateOnDEGEN(message).catch(e => {
+			await MessageCreateOnDEGEN(message).catch(e => {
 				LogUtils.logError('DEGEN failed to say hello', e);
 			});
 			// Run for webhook
 			await messageCreateOnBountyBoard(message).catch(e => {
 				LogUtils.logError('failed to create bounty message from webhook', e);
 			});
-			if (message.channel.type === 'DM') {
-				// Run scoap squad DM flow
-				await messageSetScoapRoles(message).catch(e => {
-					LogUtils.logError('failed to run scoap-squad DM flow', e);
-				});
-
-				// Run first-quest DM flow
-				await messageLaunchFirstQuest(message).catch(e => {
-					LogUtils.logError('failed to run first-quest DM flow', e);
-				});
-			}
+			// TODO: enable scoap squad at a future date
+			// if (message.channel.type === 'DM') {
+			// 	// Run scoap squad DM flow
+			// 	await messageSetScoapRoles(message).catch(e => {
+			// 		LogUtils.logError('failed to run scoap-squad DM flow', e);
+			// 	});
+			// }
 			// Check mentions for AFK users
 			if (message.mentions.users.size > 0) {
 				await HandleAFK(message).catch((e) => {