+++ src/app/service/constants/channelIds.ts
@@ -6,4 +6,5 @@ export default Object.freeze({
 	WRITERS_ROOM: process.env.DISCORD_CHANNEL_WRITERS_ROOM_ID,
 	generalSupport: process.env.DISCORD_CHANNEL_SUPPORT_ID,
 	firstQuestProject: process.env.DISCORD_CHANNEL_FIRST_QUEST_PROJECT_ID,
+	captchaVerification: process.env.DISCORD_CHANNEL_CAPTCHA_VERIFICATION,
 });