+++ src/app/utils/ServiceUtils.ts
@@ -16,11 +16,11 @@ import {
 	VoiceChannel,
 } from 'discord.js';
 import { Db } from 'mongodb';
-import { CommandContext } from 'slash-create';
+import { ButtonStyle, CommandContext, ComponentActionRow, ComponentType } from 'slash-create';
 import client from '../app';
 import ValidationError from '../errors/ValidationError';
 import constants from '../service/constants/constants';
-import roleIDs from '../service/constants/roleIds';
+import roleIds from '../service/constants/roleIds';
 import { Allowlist } from '../types/discord/Allowlist';
 import { Confusables } from './Confusables';
 import discordServerIds from '../service/constants/discordServerIds';
@@ -54,13 +54,13 @@ const ServiceUtils = {
 
 	getGuestRole(roles: RoleManager): Role {
 		return roles.cache.find((role) => {
-			return role.id === roleIDs.guestPass;
+			return role.id === roleIds.guestPass;
 		});
 	},
 
 	getAFKRole(roles: RoleManager): Role {
 		return roles.cache.find((role) => {
-			return role.id === roleIDs.AFK;
+			return role.id === roleIds.AFK;
 		});
 	},
 	
@@ -93,31 +93,51 @@ const ServiceUtils = {
 
 	isAnyLevel(guildMember: GuildMember): boolean {
 		return ServiceUtils.hasSomeRole(guildMember, [
-			roleIDs.level1,
-			roleIDs.level2,
-			roleIDs.level3,
-			roleIDs.level4,
+			roleIds.level1,
+			roleIds.level2,
+			roleIds.level3,
+			roleIds.level4,
 		]);
 	},
 
 	isAtLeastLevel1(guildMember: GuildMember): boolean {
 		return ServiceUtils.hasSomeRole(guildMember, [
-			roleIDs.level1,
-			roleIDs.level2,
-			roleIDs.level3,
-			roleIDs.level4,
-			roleIDs.admin,
-			roleIDs.genesisSquad,
+			roleIds.level1,
+			roleIds.level2,
+			roleIds.level3,
+			roleIds.level4,
+			roleIds.admin,
+			roleIds.genesisSquad,
 		]);
+	},
+	
+	/**
+	 * Level 2 members have the highest precedence among the DAO
+	 * @param guildMember
+	 */
+	isALevel2Contributor(guildMember: GuildMember): boolean {
+		return ServiceUtils.hasRole(guildMember, roleIds.level2);
+	},
+	
+	isJustAMember(guildMember: GuildMember): boolean {
+		return ServiceUtils.hasSomeRole(guildMember, [
+			roleIds.level1,
+			roleIds.level3,
+			roleIds.level4,
+		]);
+	},
+	
+	isAGuest(guildMember: GuildMember): boolean {
+		return ServiceUtils.hasRole(guildMember, roleIds.guestPass);
 	},
 
 	isAtLeastLevel2(guildMember: GuildMember): boolean {
 		return ServiceUtils.hasSomeRole(guildMember, [
-			roleIDs.level2,
-			roleIDs.level3,
-			roleIDs.level4,
-			roleIDs.admin,
-			roleIDs.genesisSquad,
+			roleIds.level2,
+			roleIds.level3,
+			roleIds.level4,
+			roleIds.admin,
+			roleIds.genesisSquad,
 		]);
 	},
 	
@@ -172,7 +192,7 @@ const ServiceUtils = {
 		}
 
 		const highRankingMembers = await ServiceUtils.getMembersWithRoles(member.guild,
-			[roleIDs.genesisSquad, roleIDs.admin, roleIDs.level2]);
+			[roleIds.genesisSquad, roleIds.admin, roleIds.level2]);
 
 		// Sanitize high-ranking member names in preparation for comparing them to new member nickname
 		const highRankingNames = highRankingMembers.map(highRankingMember => {
@@ -294,6 +314,23 @@ const ServiceUtils = {
 			errors: ['time'],
 		};
 		return (await dmChannel.awaitMessages(replyOptions)).first().attachments.first();
+	},
+	
+	sendOutErrorMessage: async (ctx: CommandContext): Promise<any> => {
+		const row: ComponentActionRow = {
+			type: ComponentType.ACTION_ROW,
+			components: [{
+				type: ComponentType.BUTTON,
+				style: ButtonStyle.LINK,
+				label: 'Support',
+				url: 'https://discord.gg/NRj43H83nJ',
+			}],
+		};
+		await ctx.send({
+			content: 'Something is not working. Please reach out to us and a support member will happily assist!',
+			ephemeral: true,
+			components: [row],
+		});
 	},
 };